requirements.txt
certifi==2024.8.30
frozendict==2.3.10
moralis==0.1.49
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
